added if conditions checking on migration fix

+++ database/migrations/2022_06_03_073702_add_columns_to_sales_orders_table.php
@@ -8,17 +8,29 @@ class AddColumnsToSalesOrdersTable extends Migration
 {
     public function up()
     {
-        Schema::table('sales_orders', function (Blueprint $table) {
-            $table->string('select_time')->nullable();
-            $table->string('vehicle_type')->nullable();
-        });
+        if (!Schema::hasColumn('sales_orders', 'select_time')) {
+            Schema::table('sales_orders', function (Blueprint $table) {
+                $table->string('select_time')->nullable();
+            });
+        }
+        if (!Schema::hasColumn('sales_orders', 'vehicle_type')) {
+            Schema::table('sales_orders', function (Blueprint $table) {
+                $table->string('vehicle_type')->nullable();
+            });
+        }
     }
 
     public function down()
     {
-        Schema::table('sales_orders', function (Blueprint $table) {
-            $table->dropColumn('select_time');
-            $table->dropColumn('vehicle_type');
-        });
+        if (Schema::hasColumn('sales_orders', 'select_time')) {
+            Schema::table('sales_orders', function (Blueprint $table) {
+                $table->dropColumn('select_time');
+            });
+        }
+        if (Schema::hasColumn('sales_orders', 'vehicle_type')) {
+            Schema::table('sales_orders', function (Blueprint $table) {
+                $table->dropColumn('vehicle_type');
+            });
+        }
     }
 }
